requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
h11==0.13.0
idna==3.3
numpy==1.23.2
outcome==1.2.0
pandas==1.4.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
xlwt==1.3.0
